fix(mobile): wire useFolderTree / FOLDER_TREE_QUERY_KEY into document screens (closes #1532)
Follow-up to PR #1464's review. The useFolderTree hook + FOLDER_TREE_QUERY_KEY
constant shipped with zero consumers — the three sites that should share the
key still hard-coded ['documents','folders','tree'], so the cache-sharing the
PR advertised was never realized and the pin-test guarded a constant nothing
depended on.

- DocumentsScreen: replace the inline useApiQuery(['documents','folders','tree'],
  '/api/v1/documents/folders/tree', {staleTime}) with useFolderTree() (uses its
  folderTree accessor). Drops the duplicated endpoint path + cache options.
- MoveDocumentSheet / NewFolderSheet: invalidate with FOLDER_TREE_QUERY_KEY
  instead of the hard-coded literal, so the constant is now load-bearing and
  the pin-test is meaningful.
- Break the reverse coupling (finding 3): ApiFolderTreeNode now lives in the
  useFolderTree hook (the data layer it describes); DocumentsScreen re-exports
  it so existing './DocumentsScreen' importers (barrel, sheets, tests) are
  unaffected. The hook no longer imports a screen module.

No behavior change — same query key, endpoint and staleTime.
